UI almost finished, function now accepts cos and sin

+++ src/Model/Sierpinski.java
@@ -145,7 +145,7 @@ public class Sierpinski extends Fractal{
         writer.close();
     }
 
-    public int[][] getTab() {
+    public double[][] getTab() {
         return tab;
     }
 }
